Address comments - 3 | Major refactor
