+++ timelock-corruption-detection/build.gradle
@@ -15,7 +15,6 @@ dependencies {
 
     compile project(":timelock-corruption-detection:timelock-corruption-detection-undertow")
     compile project(":timelock-corruption-detection:timelock-corruption-detection-jersey")
-    compile project(":timelock-corruption-detection:timelock-corruption-detection-objects")
 
     compile group: 'com.google.guava', name: 'guava'
     compile group: 'com.palantir.conjure.java.api', name: 'service-config'
