+++ timelock-corruption-detection/src/main/java/com/palantir/timelock/history/LocalHistoryLoader.java
@@ -14,19 +14,19 @@
  * limitations under the License.
  */
 
-package com.palantir.history;
+package com.palantir.timelock.history;
 
 import java.util.Map;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.palantir.common.streams.KeyedStream;
-import com.palantir.history.models.AcceptorUseCase;
-import com.palantir.history.models.ImmutablePaxosHistoryOnSingleNode;
-import com.palantir.history.models.LearnerAndAcceptorRecords;
-import com.palantir.history.models.LearnerUseCase;
-import com.palantir.history.models.PaxosHistoryOnSingleNode;
-import com.palantir.history.sqlite.SqlitePaxosStateLogHistory;
 import com.palantir.paxos.NamespaceAndUseCase;
+import com.palantir.timelock.history.models.AcceptorUseCase;
+import com.palantir.timelock.history.models.ImmutablePaxosHistoryOnSingleNode;
+import com.palantir.timelock.history.models.LearnerAndAcceptorRecords;
+import com.palantir.timelock.history.models.LearnerUseCase;
+import com.palantir.timelock.history.models.PaxosHistoryOnSingleNode;
+import com.palantir.timelock.history.sqlite.SqlitePaxosStateLogHistory;
 
 //TBD cache implementation
 public final class LocalHistoryLoader {
