+++ timelock-corruption-detection/src/main/java/com/palantir/timelock/history/models/LearnerAndAcceptorRecords.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.history.models;
+package com.palantir.timelock.history.models;
 
 import java.util.Comparator;
 import java.util.Map;
