+++ timelock-corruption-detection/src/main/java/com/palantir/timelock/history/sqlite/LogVerificationProgressState.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.history.sqlite;
+package com.palantir.timelock.history.sqlite;
 
 import java.util.OptionalLong;
 import java.util.function.Function;
