+++ timelock-corruption-detection/src/test/java/com/palantir/history/util/UseCaseUtilsTest.java
@@ -22,6 +22,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.Test;
 
 import com.palantir.logsafe.exceptions.SafeIllegalStateException;
+import com.palantir.timelock.history.util.UseCaseUtils;
 
 public class UseCaseUtilsTest {
     @Test
