+++ timelock-corruption-detection/src/test/java/com/palantir/timelock/history/SqliteHistoryQueriesTest.java
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package com.palantir.history;
+package com.palantir.timelock.history;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import static com.palantir.history.utils.PaxosSerializationTestUtils.writeAcceptorStateForLogAndRound;
-import static com.palantir.history.utils.PaxosSerializationTestUtils.writeValueForLogAndRound;
+import static com.palantir.paxos.PaxosStateLogTestUtils.longToBytes;
+import static com.palantir.timelock.history.utils.PaxosSerializationTestUtils.writeAcceptorStateForLogAndRound;
+import static com.palantir.timelock.history.utils.PaxosSerializationTestUtils.writeValueForLogAndRound;
 
 import java.util.Set;
 import java.util.stream.IntStream;
@@ -31,9 +32,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
-import com.palantir.history.models.LearnerAndAcceptorRecords;
-import com.palantir.history.sqlite.SqlitePaxosStateLogHistory;
-import com.palantir.history.util.UseCaseUtils;
+import com.palantir.timelock.history.models.LearnerAndAcceptorRecords;
+import com.palantir.timelock.history.sqlite.SqlitePaxosStateLogHistory;
+import com.palantir.timelock.history.util.UseCaseUtils;
 import com.palantir.paxos.Client;
 import com.palantir.paxos.ImmutableNamespaceAndUseCase;
 import com.palantir.paxos.NamespaceAndUseCase;
@@ -42,6 +43,7 @@ import com.palantir.paxos.PaxosStateLog;
 import com.palantir.paxos.PaxosValue;
 import com.palantir.paxos.SqliteConnections;
 import com.palantir.paxos.SqlitePaxosStateLog;
+import com.palantir.timelock.history.utils.PaxosSerializationTestUtils;
 
 public class SqliteHistoryQueriesTest {
     @Rule
